+++ plugin-packs/postcss-preset-env/src/lib/get-unsupported-browsers-by-feature.mjs
@@ -1,26 +1,7 @@
-const browsers = [
-	'and_chr',
-	'and_ff',
-	'and_qq',
-	'and_uc',
-	'android',
-	'baidu',
-	'chrome',
-	'edge',
-	'firefox',
-	'ie',
-	'ie_mob',
-	'ios_saf',
-	'kaios',
-	'op_mini',
-	'op_mob',
-	'opera',
-	'safari',
-	'samsung',
-];
+import { browsersWithSupportStats } from './browsers-with-supports-stats.mjs';
 
 // return a list of browsers that do not support the feature
-export default function getUnsupportedBrowsersByFeature(feature) {
+export function getUnsupportedBrowsersByFeature(feature) {
 	if (!feature) {
 		return [];
 	}
@@ -32,7 +13,7 @@ export default function getUnsupportedBrowsersByFeature(feature) {
 
 	const query = [];
 
-	browsers.forEach(browser => {
+	browsersWithSupportStats.forEach(browser => {
 		if (browser === 'op_mini' && (typeof feature.browser_support[browser] === 'undefined')) {
 			// Opera Mini is always "op_mini all"
 			query.push('op_mini all');